[hal] Sync: Move multiplexing responsibility to the top level.
In this commit, I remove the responsibility of HAL to handle the
multiplexing of NoC resources for each communication service. In
this way, the exclusive allocation of a resource is guaranteed
throughout the lifetime of the service exposed by HAL.

Another important point is the creation of the implementation and
definition checks of the NoC functions and variables.

Related Issues: #51, #91, #92, #93, #94, #95, #96 and #97.