Bug Fix: Unmask Clock Interrupt
Previously, we were assuming that all interrupt masks were cleared, so
that if we simply enabled hardware interrupts clock interrupts would
follow. However, enabling hardware interrupts and unmasking them are two
different things that we were missing to do. In this commit, I ensure
that we unmask the clock interrupt line before proceeding with tests.