[hal] Enhancement: MMIO Interface
Since HAL must take care of the entire MMIO structure (rather than the
architecture), it does not make sense to keep the flags that indicate
whether an arch supports MMIO or not.

In addition, the interface has been drastically simplified and all read
and write functions have been removed, since the only function that HAL
should handle are the appropriate addresses for writing to the devices.

For this, a function called mmio_get() was created, which returns the
appropriate address for the target, according to the current state of
the MMU. Therefore, the only requirement that all targets should
implement is the mmu_is_enabled() function, that indicates the current
state of the MMU. For architectures that have a constantly active MMU,
it is recommended to only return a positive value, other than zero.