Enhancement: Auto Mask/Unmask Interrupts
Previously, after registering/unregistering an interrupt handler for a
given interrupt, the interrupt would need to be masked/unmasked by
explicitly calling either interrupt_mask() or interrupt_unmask(). In
this commit, I make things more natural: the interrupt is automatically
masked/unmasked when a handler is registered/unregistered for it.