Bug Fix: Bad Interrupt Counting in Unit Tests
Previously, we were badly counting the number of interrupts in the
"Enable and Disable Interrupts" unit tests. To put it simply, we were
note ensuring that we were awaiting for a call to the dummy interrupt
handler, and thus that an interrupt had been triggered and handled. In
this commit, I fix this unit test to behave as it was designed.