Bug Fix: Dropping Routines from HW Int. Interface
Currently, the Hardware Interrupt Interface exposes routines for masking
and unmasking interrupts, as well as raising and dropping interrupt
levels.

Although this enables maximum flexibility the current interface is not
so well designed to prevent some bad usage of the interface, such as
unmasking interrupts before registering an interrupt handler.

While we do not solve these problems, we should not expose the following
routines of the Hardware Interrupt Interface:

- interrupt_set_handler()
- interrupt_mask()
- interrupt_unmaks()
- interrupt_set_level()