[hal] Bug Fix: Optimization Produces Wrong Code
The GCC was producing wrong code if the optimization flags
were enabled, because the condition variables were changed
in other contexts other than the functions they were in,
so the compiler was producing an infinite loop instead of
checking the value of the variable at all times.

To fix, the 'volatile' qualifier is needed, to tell the
compiler that the variable can be changed in other contexts.