[k1b-core] Bugfix: Minimize the starvation on spinlock.
Insert a silly job in the lock to minimize competition and the
starvation problem in the spinlock.

As I was able to recover the execution once using gdb (but it blocked
some seconds later), I believe it is a starvation that comes from the
competition of the instruction that bypasses the cache. Because if it
were the loss of the release value
(1ULL), recovery would be impossible.